+++ src/Parameters.cpp
@@ -166,15 +166,15 @@ ezc3d::ParametersNS::Parameters::Parameters(ezc3d::c3d &file) :
     _nbParamBlock = file.readInt(1*ezc3d::DATA_TYPE::BYTE);
     _processorType = file.readInt(1*ezc3d::DATA_TYPE::BYTE);
     if (_checksum == 0 && _parametersStart == 0){
-        // Theoritically, this is a bad c3d formatting and should return an error, but for some reason
-        // qualisys decided that they would not comply to the standard a put _parameterStart and _checksum to 0
+        // Theoritically, if this happens, this is a bad c3d formatting and should return an error, but for some reason
+        // Qualisys decided that they would not comply to the standard. Therefore they put "_parameterStart" and "_checksum" to 0
         // This is a patch for Qualisys bad formatting c3d
         _parametersStart = 1;
         _checksum = 0x50;
     }
 
     // Read parameter or group
-    std::streampos nextParamByteInFile((int)file.tellg() + _parametersStart - ezc3d::DATA_TYPE::BYTE);
+    std::streampos nextParamByteInFile(static_cast<int>(file.tellg()) + _parametersStart - ezc3d::DATA_TYPE::BYTE);
     while (nextParamByteInFile)
     {
         // Check if we spontaneously got to the next parameter. Otherwise c3d is messed up
@@ -188,7 +188,7 @@ ezc3d::ParametersNS::Parameters::Parameters(ezc3d::c3d &file) :
         int id(file.readInt(1*ezc3d::DATA_TYPE::BYTE));
 
         // Make sure there at least enough group
-        for (size_t i = _groups.size(); i < abs(id); ++i)
+        for (int i = static_cast<int>(_groups.size()); i < abs(id); ++i)
             _groups.push_back(ezc3d::ParametersNS::GroupNS::Group());
 
         // Group ID always negative for groups and positive parameter of group ID
@@ -221,7 +221,7 @@ void ezc3d::ParametersNS::Parameters::print() const
     std::cout << "nbParamBlock = " << nbParamBlock() << std::endl;
     std::cout << "processorType = " << processorType() << std::endl;
 
-    for (int i = 0; i < groups().size(); ++i){
+    for (int i = 0; i < static_cast<int>(groups().size()); ++i){
         std::cout << "Group " << i << std::endl;
         group(i).print();
         std::cout << std::endl;
@@ -244,13 +244,13 @@ void ezc3d::ParametersNS::Parameters::write(std::fstream &f) const
 
     // Write each groups
     std::streampos dataStartPosition; // Special parameter in POINT group
-    for (int i=0; i<groups().size(); ++i){
+    for (int i=0; i<static_cast<int>(groups().size()); ++i){
         group(i).write(f, -(i+1), dataStartPosition);
     }
 
     // Move the cursor to a beginning of a block
     std::streampos actualPos(f.tellg());
-    for (int i=0; i<512 - (int)actualPos % 512; ++i){
+    for (int i=0; i<512 - static_cast<int>(actualPos) % 512; ++i){
         f.write(reinterpret_cast<const char*>(&blankValue), ezc3d::BYTE);
     }
     // Go back at the left blank space and write the actual position
@@ -288,7 +288,7 @@ const std::vector<ezc3d::ParametersNS::GroupNS::Group>& ezc3d::ParametersNS::Par
 }
 ezc3d::ParametersNS::GroupNS::Group &ezc3d::ParametersNS::Parameters::group_nonConst(int group)
 {
-    return _groups[group];
+    return _groups[static_cast<unsigned int>(group)];
 }
 
 void ezc3d::ParametersNS::Parameters::addGroup(const ezc3d::ParametersNS::GroupNS::Group &g)
@@ -298,14 +298,14 @@ void ezc3d::ParametersNS::Parameters::addGroup(const ezc3d::ParametersNS::GroupN
     if (alreadyExtIdx < 0)
         _groups.push_back(g);
     else {
-        for (int i=0; i<g.parameters().size(); ++i)
-            _groups[alreadyExtIdx].addParameter(g.parameter(i));
+        for (int i=0; i<static_cast<int>(g.parameters().size()); ++i)
+            _groups[static_cast<unsigned int>(alreadyExtIdx)].addParameter(g.parameter(i));
     }
 
 }
 const ezc3d::ParametersNS::GroupNS::Group &ezc3d::ParametersNS::Parameters::group(int group) const
 {
-    return _groups[group];
+    return _groups[static_cast<unsigned int>(group)];
 }
 const ezc3d::ParametersNS::GroupNS::Group &ezc3d::ParametersNS::Parameters::group(const std::string &groupName) const
 {
@@ -317,7 +317,7 @@ const ezc3d::ParametersNS::GroupNS::Group &ezc3d::ParametersNS::Parameters::grou
 
 int ezc3d::ParametersNS::Parameters::groupIdx(const std::string &groupName) const
 {
-    for (int i = 0; i < groups().size(); ++i){
+    for (int i = 0; i < static_cast<int>(groups().size()); ++i){
         if (!group(i).name().compare(groupName))
             return i;
     }
@@ -351,22 +351,22 @@ int ezc3d::ParametersNS::GroupNS::Group::read(ezc3d::c3d &file, int nbCharInName
         _isLocked = false;
 
     // Read name of the group
-    _name.assign(file.readString(abs(nbCharInName) * ezc3d::DATA_TYPE::BYTE));
+    _name.assign(file.readString(static_cast<unsigned int>(abs(nbCharInName) * ezc3d::DATA_TYPE::BYTE)));
 
     // number of byte to the next group from here
-    int offsetNext((int)file.readUint(2*ezc3d::DATA_TYPE::BYTE));
+    int offsetNext(static_cast<int>(file.readUint(2*ezc3d::DATA_TYPE::BYTE)));
     // Compute the position of the element in the file
     int nextParamByteInFile;
     if (offsetNext == 0)
         nextParamByteInFile = 0;
     else
-        nextParamByteInFile = (int)file.tellg() + offsetNext - ezc3d::DATA_TYPE::WORD;
+        nextParamByteInFile = static_cast<int>(file.tellg()) + offsetNext - ezc3d::DATA_TYPE::WORD;
 
     // Byte 5+nbCharInName ==> Number of characters in group description
     int nbCharInDesc(file.readInt(1*ezc3d::DATA_TYPE::BYTE));
     // Byte 6+nbCharInName ==> Group description
     if (nbCharInDesc)
-        _description = file.readString(nbCharInDesc);
+        _description = file.readString(static_cast<unsigned int>(nbCharInDesc));
 
     // Return how many bytes
     return nextParamByteInFile;
@@ -385,7 +385,7 @@ void ezc3d::ParametersNS::GroupNS::Group::addParameter(const ezc3d::ParametersNS
         throw std::runtime_error("Data type is not set");
 
     int alreadyExistIdx(-1);
-    for (int i=0; i<_parameters.size(); ++i)
+    for (int i=0; i<static_cast<int>(_parameters.size()); ++i)
         if (!parameter(i).name().compare(p.name())){
             alreadyExistIdx = i;
             break;
@@ -393,7 +393,7 @@ void ezc3d::ParametersNS::GroupNS::Group::addParameter(const ezc3d::ParametersNS
     if (alreadyExistIdx < 0)
         _parameters.push_back(p);
     else
-        _parameters[alreadyExistIdx] = p;
+        _parameters[static_cast<unsigned int>(alreadyExistIdx)] = p;
 }
 void ezc3d::ParametersNS::GroupNS::Group::print() const
 {
@@ -401,14 +401,14 @@ void ezc3d::ParametersNS::GroupNS::Group::print() const
     std::cout << "isLocked = " << isLocked() << std::endl;
     std::cout << "desc = " << description() << std::endl;
 
-    for (int i=0; i<parameters().size(); ++i){
+    for (int i=0; i<static_cast<int>(parameters().size()); ++i){
         std::cout << "Parameter " << i << std::endl;
         parameter(i).print();
     }
 }
 void ezc3d::ParametersNS::GroupNS::Group::write(std::fstream &f, int groupIdx, std::streampos &dataStartPosition) const
 {
-    int nCharName(name().size());
+    int nCharName(static_cast<int>(name().size()));
     if (isLocked())
         nCharName *= -1;
     f.write(reinterpret_cast<const char*>(&nCharName), 1*ezc3d::DATA_TYPE::BYTE);
@@ -422,7 +422,7 @@ void ezc3d::ParametersNS::GroupNS::Group::write(std::fstream &f, int groupIdx, s
     std::streampos pos(f.tellg());
     f.write(reinterpret_cast<const char*>(&blank), 2*ezc3d::DATA_TYPE::BYTE);
 
-    int nCharGroupDescription(description().size());
+    int nCharGroupDescription(static_cast<int>(description().size()));
     f.write(reinterpret_cast<const char*>(&nCharGroupDescription), 1*ezc3d::DATA_TYPE::BYTE);
     f.write(description().c_str(), nCharGroupDescription*ezc3d::DATA_TYPE::BYTE);
 
@@ -433,7 +433,7 @@ void ezc3d::ParametersNS::GroupNS::Group::write(std::fstream &f, int groupIdx, s
     f.write(reinterpret_cast<const char*>(&nCharToNext), 2*ezc3d::DATA_TYPE::BYTE);
     f.seekg(actualPos);
 
-    for (int i=0; i<parameters().size(); ++i){
+    for (int i=0; i<static_cast<int>(parameters().size()); ++i){
         parameter(i).write(f, -groupIdx, dataStartPosition);
     }
 }
@@ -444,9 +444,9 @@ const std::vector<ezc3d::ParametersNS::GroupNS::Parameter>& ezc3d::ParametersNS:
 
 const ezc3d::ParametersNS::GroupNS::Parameter &ezc3d::ParametersNS::GroupNS::Group::parameter(int idx) const
 {
-    if (idx < 0 || idx >= _parameters.size())
+    if (idx < 0 || idx >= static_cast<int>(_parameters.size()))
         throw std::out_of_range("Wrong number of parameter");
-    return _parameters[idx];
+    return _parameters[static_cast<unsigned int>(idx)];
 }
 std::vector<ezc3d::ParametersNS::GroupNS::Parameter>& ezc3d::ParametersNS::GroupNS::Group::parameters_nonConst()
 {
@@ -454,7 +454,7 @@ std::vector<ezc3d::ParametersNS::GroupNS::Parameter>& ezc3d::ParametersNS::Group
 }
 int ezc3d::ParametersNS::GroupNS::Group::parameterIdx(std::string parameterName) const
 {
-    for (int i = 0; i < parameters().size(); ++i){
+    for (int i = 0; i < static_cast<int>(parameters().size()); ++i){
         if (!parameter(i).name().compare(parameterName))
             return i;
     }
@@ -512,16 +512,16 @@ int ezc3d::ParametersNS::GroupNS::Parameter::read(ezc3d::c3d &file, int nbCharIn
         _isLocked = false;
 
     // Read name of the group
-    _name = file.readString(abs(nbCharInName) * ezc3d::DATA_TYPE::BYTE);
+    _name = file.readString(static_cast<unsigned int>(abs(nbCharInName) * ezc3d::DATA_TYPE::BYTE));
 
     // number of byte to the next group from here
-    int offsetNext((int)file.readUint(2*ezc3d::DATA_TYPE::BYTE));
+    int offsetNext(static_cast<int>(file.readUint(2*ezc3d::DATA_TYPE::BYTE)));
     // Compute the position of the element in the file
     int nextParamByteInFile;
     if (offsetNext == 0)
         nextParamByteInFile = 0;
     else
-        nextParamByteInFile = (int)file.tellg() + offsetNext - ezc3d::DATA_TYPE::WORD;
+        nextParamByteInFile = static_cast<int>(file.tellg()) + offsetNext - ezc3d::DATA_TYPE::WORD;
 
     // -1 sizeof(char), 1 byte, 2 int, 4 float
     int lengthInByte(file.readInt(1*ezc3d::DATA_TYPE::BYTE));
@@ -548,9 +548,9 @@ int ezc3d::ParametersNS::GroupNS::Parameter::read(ezc3d::c3d &file, int nbCharIn
     if (_data_type == DATA_TYPE::CHAR)
         file.readMatrix(_dimension, _param_data_string);
     else if (_data_type == DATA_TYPE::BYTE)
-        file.readMatrix((int)_data_type, _dimension, _param_data_int);
+        file.readMatrix(static_cast<unsigned int>(_data_type), _dimension, _param_data_int);
     else if (_data_type == DATA_TYPE::INT)
-        file.readMatrix((int)_data_type, _dimension, _param_data_int);
+        file.readMatrix(static_cast<unsigned int>(_data_type), _dimension, _param_data_int);
     else if (_data_type == DATA_TYPE::FLOAT)
         file.readMatrix(_dimension, _param_data_float);
 
@@ -559,7 +559,7 @@ int ezc3d::ParametersNS::GroupNS::Parameter::read(ezc3d::c3d &file, int nbCharIn
     int nbCharInDesc(file.readInt(1*ezc3d::DATA_TYPE::BYTE));
     // Byte 6+nbCharInName ==> Group description
     if (nbCharInDesc)
-        _description = file.readString(nbCharInDesc);
+        _description = file.readString(static_cast<unsigned int>(nbCharInDesc));
 
     // Return how many bytes
     return nextParamByteInFile;
@@ -568,7 +568,7 @@ int ezc3d::ParametersNS::GroupNS::Parameter::read(ezc3d::c3d &file, int nbCharIn
 bool ezc3d::ParametersNS::GroupNS::Parameter::isDimensionConsistent(int dataSize, const std::vector<int>& dimension) const {
     if (dataSize == 0){
         int dim(1);
-        for (int i=0; i<dimension.size(); ++i)
+        for (unsigned int i=0; i<dimension.size(); ++i)
             dim *= dimension[i];
         if (dimension.size() == 0 || dim == 0)
             return true;
@@ -577,7 +577,7 @@ bool ezc3d::ParametersNS::GroupNS::Parameter::isDimensionConsistent(int dataSize
     }
 
     int dimesionSize(1);
-    for (int i=0; i<dimension.size(); ++i)
+    for (unsigned int i=0; i<dimension.size(); ++i)
         dimesionSize *= dimension[i];
     if (dataSize == dimesionSize)
         return true;
@@ -586,7 +586,7 @@ bool ezc3d::ParametersNS::GroupNS::Parameter::isDimensionConsistent(int dataSize
 }
 void ezc3d::ParametersNS::GroupNS::Parameter::set(const std::vector<int> &data, const std::vector<int>& dimension)
 {
-    if (!isDimensionConsistent(data.size(), dimension))
+    if (!isDimensionConsistent(static_cast<int>(data.size()), dimension))
         throw std::range_error("Dimension of the data does not correspond to sent dimensions");
     _data_type = ezc3d::DATA_TYPE::INT;
     _param_data_int = data;
@@ -594,7 +594,7 @@ void ezc3d::ParametersNS::GroupNS::Parameter::set(const std::vector<int> &data, 
 }
 void ezc3d::ParametersNS::GroupNS::Parameter::set(const std::vector<float> &data, const std::vector<int> &dimension)
 {
-    if (!isDimensionConsistent(data.size(), dimension))
+    if (!isDimensionConsistent(static_cast<int>(data.size()), dimension))
         throw std::range_error("Dimension of the data does not correspond to sent dimensions");
     _data_type = ezc3d::DATA_TYPE::FLOAT;
     _param_data_float = data;
@@ -603,9 +603,9 @@ void ezc3d::ParametersNS::GroupNS::Parameter::set(const std::vector<float> &data
 void ezc3d::ParametersNS::GroupNS::Parameter::set(const std::vector<std::string> &data, const std::vector<int> &dimension)
 {
     std::vector<int> dimensionNoStr;
-    for (int i=1; i<dimension.size(); ++i)
+    for (unsigned int i=1; i<dimension.size(); ++i)
         dimensionNoStr.push_back(dimension[i]);
-    if (!isDimensionConsistent(data.size(), dimensionNoStr))
+    if (!isDimensionConsistent(static_cast<int>(data.size()), dimensionNoStr))
         throw std::range_error("Dimension of the data does not correspond to sent dimensions");
     _data_type = ezc3d::DATA_TYPE::CHAR;
     _param_data_string = data;
@@ -619,16 +619,16 @@ void ezc3d::ParametersNS::GroupNS::Parameter::print() const
 
     // Data are not separated according to _dimension, which could help to read
     if (_data_type == DATA_TYPE::CHAR)
-        for (int i = 0; i < _param_data_string.size(); ++i)
+        for (unsigned int i = 0; i < _param_data_string.size(); ++i)
             std::cout << "param_data_string[" << i << "] = " << _param_data_string[i] << std::endl;
     if (_data_type == DATA_TYPE::BYTE)
-        for (int i = 0; i < _param_data_int.size(); ++i)
+        for (unsigned int i = 0; i < _param_data_int.size(); ++i)
             std::cout << "param_data[" << i << "] = " << _param_data_int[i] << std::endl;
     if (_data_type == DATA_TYPE::INT)
-        for (int i = 0; i < _param_data_int.size(); ++i)
+        for (unsigned int i = 0; i < _param_data_int.size(); ++i)
             std::cout << "param_data[" << i << "] = " << _param_data_int[i] << std::endl;
     if (_data_type == DATA_TYPE::FLOAT)
-        for (int i = 0; i < _param_data_float.size(); ++i)
+        for (unsigned int i = 0; i < _param_data_float.size(); ++i)
             std::cout << "param_data[" << i << "] = " << _param_data_float[i] << std::endl;
 
     std::cout << "description = " << _description << std::endl;
@@ -636,7 +636,7 @@ void ezc3d::ParametersNS::GroupNS::Parameter::print() const
 
 void ezc3d::ParametersNS::GroupNS::Parameter::write(std::fstream &f, int groupIdx, std::streampos &dataStartPosition) const
 {
-    int nCharName(name().size());
+    int nCharName(static_cast<int>(name().size()));
     if (isLocked())
         nCharName *= -1;
     f.write(reinterpret_cast<const char*>(&nCharName), 1*ezc3d::DATA_TYPE::BYTE);
@@ -652,7 +652,7 @@ void ezc3d::ParametersNS::GroupNS::Parameter::write(std::fstream &f, int groupId
 
     // Write the parameter values
     f.write(reinterpret_cast<const char*>(&_data_type), 1*ezc3d::DATA_TYPE::BYTE);
-    int size_dim(_dimension.size());
+    int size_dim(static_cast<int>(_dimension.size()));
     // If it is a scalar, store it as so
     if (_dimension.size() == 1 && _dimension[0] == 1){
         int _size_dim(0);
@@ -660,16 +660,16 @@ void ezc3d::ParametersNS::GroupNS::Parameter::write(std::fstream &f, int groupId
     }
     else{
         f.write(reinterpret_cast<const char*>(&size_dim), 1*ezc3d::DATA_TYPE::BYTE);
-        for (int i=0; i<_dimension.size(); ++i)
+        for (unsigned int i=0; i<_dimension.size(); ++i)
             f.write(reinterpret_cast<const char*>(&_dimension[i]), 1*ezc3d::DATA_TYPE::BYTE);
     }
     int hasSize(1);
-    for (int i=0; i<_dimension.size(); ++i)
+    for (unsigned int i=0; i<_dimension.size(); ++i)
         hasSize *= _dimension[i];
     if (hasSize > 0){
         if (_data_type == DATA_TYPE::CHAR){
             if (_dimension.size() == 1){
-                f.write(_param_data_string[0].c_str(), _param_data_string[0].size()*(int)DATA_TYPE::BYTE);
+                f.write(_param_data_string[0].c_str(), static_cast<int>(_param_data_string[0].size())*static_cast<int>(DATA_TYPE::BYTE));
             } else {
                 writeImbricatedParameter(f, _dimension, 1);
             }
@@ -684,7 +684,7 @@ void ezc3d::ParametersNS::GroupNS::Parameter::write(std::fstream &f, int groupId
     }
 
     // Write description of the parameter
-    int nCharDescription(description().size());
+    int nCharDescription(static_cast<int>(description().size()));
     f.write(reinterpret_cast<const char*>(&nCharDescription), 1*ezc3d::DATA_TYPE::BYTE);
     f.write(description().c_str(), nCharDescription*ezc3d::DATA_TYPE::BYTE);
 
@@ -700,20 +700,20 @@ const std::vector<int> ezc3d::ParametersNS::GroupNS::Parameter::dimension() cons
 {
     return _dimension;
 }
-int ezc3d::ParametersNS::GroupNS::Parameter::writeImbricatedParameter(std::fstream &f, const std::vector<int>& dim, int currentIdx, int cmp) const{
+unsigned int ezc3d::ParametersNS::GroupNS::Parameter::writeImbricatedParameter(std::fstream &f, const std::vector<int>& dim, unsigned int currentIdx, unsigned int cmp) const{
     for (int i=0; i<dim[currentIdx]; ++i)
         if (currentIdx == dim.size()-1){
             if (_data_type == DATA_TYPE::BYTE)
-                f.write(reinterpret_cast<const char*>(&(_param_data_int[cmp])), (int)_data_type);
+                f.write(reinterpret_cast<const char*>(&(_param_data_int[cmp])), static_cast<int>(_data_type));
             else if (_data_type == DATA_TYPE::INT)
-                f.write(reinterpret_cast<const char*>(&(_param_data_int[cmp])), (int)_data_type);
+                f.write(reinterpret_cast<const char*>(&(_param_data_int[cmp])), static_cast<int>(_data_type));
             else if (_data_type == DATA_TYPE::FLOAT)
-                f.write(reinterpret_cast<const char*>(&(_param_data_float[cmp])), (int)_data_type);
+                f.write(reinterpret_cast<const char*>(&(_param_data_float[cmp])), static_cast<int>(_data_type));
             else if (_data_type == DATA_TYPE::CHAR){
-                f.write(_param_data_string[cmp].c_str(), _param_data_string[cmp].size()*(int)DATA_TYPE::BYTE);
+                f.write(_param_data_string[cmp].c_str(), static_cast<int>(_param_data_string[cmp].size())*static_cast<int>(DATA_TYPE::BYTE));
                 const char buffer = ' ';
-                for (int j=_param_data_string[cmp].size(); j<_dimension[0]; ++j)
-                    f.write(&buffer, (int)DATA_TYPE::BYTE);
+                for (int j=static_cast<int>(_param_data_string[cmp].size()); j<_dimension[0]; ++j)
+                    f.write(&buffer, static_cast<int>(DATA_TYPE::BYTE));
             }
             ++cmp;
         }
